# docs/swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "Starter API",
    "contact": {}
  },
  "host": "localhost:8080.",
  "paths": {
    "/api/v1/addresses": {
      "get": {
        "consumes": ["application/json"],
        "produces": ["application/json"],
        "summary": "Возвращает список пользователей",
        "parameters": [
          {
            "type": "string",
            "description": "Authentication header",
            "name": "Authorization",
            "in": "header",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "allOf": [
                {
                  "$ref": "#/definitions/model.ResponseSuccess"
                },
                {
                  "type": "object",
                  "properties": {
                    "data": {
                      "$ref": "#/definitions/responses.getAddressesResponseData"
                    }
                  }
                }
              ]
            }
          },
          "403": {
            "description": "Коды ошибок: [1100]",
            "schema": {
              "$ref": "#/definitions/model.ResponseError"
            }
          }
        }
      }
    },
    "/api/v1/booking": {
      "post": {
        "consumes": ["application/json"],
        "produces": ["application/json"],
        "summary": "Бронирование креативной площадки",
        "parameters": [
          {
            "description": "Параметры запроса",
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/requests.createBookingRequestData"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "allOf": [
                {
                  "$ref": "#/definitions/model.ResponseSuccess"
                },
                {
                  "type": "object",
                  "properties": {
                    "data": {
                      "$ref": "#/definitions/responses.createBookingResponseData"
                    }
                  }
                }
              ]
            }
          },
          "403": {
            "description": "Коды ошибок: [1100]",
            "schema": {
              "$ref": "#/definitions/model.ResponseError"
            }
          }
        }
      }
    },
    "/api/v1/bookings": {
      "get": {
        "consumes": ["application/json"],
        "produces": ["application/json"],
        "summary": "Возвращает список бронирований",
        "parameters": [
          {
            "type": "string",
            "description": "Authentication header",
            "name": "Authorization",
            "in": "header",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "allOf": [
                {
                  "$ref": "#/definitions/model.ResponseSuccess"
                },
                {
                  "type": "object",
                  "properties": {
                    "data": {
                      "$ref": "#/definitions/responses.getBookingsResponseData"
                    }
                  }
                }
              ]
            }
          },
          "403": {
            "description": "Коды ошибок: [1100]",
            "schema": {
              "$ref": "#/definitions/model.ResponseError"
            }
          }
        }
      }
    },
    "/api/v1/bookings/:id": {
      "get": {
        "consumes": ["application/json"],
        "produces": ["application/json"],
        "summary": "Возвращает бронирование по id",
        "parameters": [
          {
            "type": "string",
            "description": "Authentication header",
            "name": "Authorization",
            "in": "header",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "allOf": [
                {
                  "$ref": "#/definitions/model.ResponseSuccess"
                },
                {
                  "type": "object",
                  "properties": {
                    "data": {
                      "$ref": "#/definitions/responses.getBookingResponseData"
                    }
                  }
                }
              ]
            }
          },
          "403": {
            "description": "Коды ошибок: [1100]",
            "schema": {
              "$ref": "#/definitions/model.ResponseError"
            }
          }
        }
      },
      "delete": {
        "consumes": ["application/json"],
        "produces": ["application/json"],
        "summary": "Удаление информации о бронировании",
        "parameters": [
          {
            "type": "string",
            "description": "Id бронирования",
            "name": "id",
            "in": "path",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/model.ResponseSuccess"
            }
          },
          "403": {
            "description": "Коды ошибок: [1100]",
            "schema": {
              "$ref": "#/definitions/model.ResponseError"
            }
          }
        }
      },
      "patch": {
        "consumes": ["application/json"],
        "produces": ["application/json"],
        "summary": "Изменение дат бронирования креативной площадки",
        "parameters": [
          {
            "description": "Параметры запроса",
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/requests.patchBookingRequestData"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/model.ResponseSuccess"
            }
          },
          "403": {
            "description": "Коды ошибок: [1100]",
            "schema": {
              "$ref": "#/definitions/model.ResponseError"
            }
          }
        }
      }
    },
    "/api/v1/calendar/convert": {
      "post": {
        "consumes": ["multipart/form-data"],
        "produces": ["application/json"],
        "summary": "Возвращает события календаря",
        "parameters": [
          {
            "type": "string",
            "description": "Authentication header",
            "name": "Authorization",
            "in": "header",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "allOf": [
                {
                  "$ref": "#/definitions/model.ResponseSuccess"
                },
                {
                  "type": "object",
                  "properties": {
                    "data": {
                      "$ref": "#/definitions/responses.convertCalendarResponseData"
                    }
                  }
                }
              ]
            }
          },
          "403": {
            "description": "Коды ошибок: [1100]",
            "schema": {
              "$ref": "#/definitions/model.ResponseError"
            }
          }
        }
      }
    },
    "/api/v1/calendar/convertFromLink": {
      "post": {
        "consumes": ["application/json"],
        "produces": ["application/json"],
        "summary": "Возвращает события календаря",
        "parameters": [
          {
            "type": "string",
            "description": "Authentication header",
            "name": "Authorization",
            "in": "header",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "allOf": [
                {
                  "$ref": "#/definitions/model.ResponseSuccess"
                },
                {
                  "type": "object",
                  "properties": {
                    "data": {
                      "$ref": "#/definitions/responses.convertCalendarFromLinkResponseData"
                    }
                  }
                }
              ]
            }
          },
          "403": {
            "description": "Коды ошибок: [1100]",
            "schema": {
              "$ref": "#/definitions/model.ResponseError"
            }
          }
        }
      }
    },
    "/api/v1/creativeSpace": {
      "post": {
        "consumes": ["application/json"],
        "produces": ["application/json"],
        "summary": "Создание креативной площадки",
        "parameters": [
          {
            "description": "Параметры запроса",
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/requests.createCreativeSpaceRequestData"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "allOf": [
                {
                  "$ref": "#/definitions/model.ResponseSuccess"
                },
                {
                  "type": "object",
                  "properties": {
                    "data": {
                      "$ref": "#/definitions/responses.createCreativeSpaceResponseData"
                    }
                  }
                }
              ]
            }
          },
          "403": {
            "description": "Коды ошибок: [1100]",
            "schema": {
              "$ref": "#/definitions/model.ResponseError"
            }
          }
        }
      }
    },
    "/api/v1/creativeSpaces": {
      "get": {
        "consumes": ["application/json"],
        "produces": ["application/json"],
        "summary": "Возвращает список креативных площадок",
        "parameters": [
          {
            "type": "string",
            "description": "Authentication header",
            "name": "Authorization",
            "in": "header",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "allOf": [
                {
                  "$ref": "#/definitions/model.ResponseSuccess"
                },
                {
                  "type": "object",
                  "properties": {
                    "data": {
                      "$ref": "#/definitions/responses.getCreativeSpacesResponseData"
                    }
                  }
                }
              ]
            }
          },
          "403": {
            "description": "Коды ошибок: [1100]",
            "schema": {
              "$ref": "#/definitions/model.ResponseError"
            }
          }
        }
      }
    },
    "/api/v1/creativeSpaces/:id": {
      "get": {
        "produces": ["application/json"],
        "summary": "Возвращает информацию о пользователе",
        "parameters": [
          {
            "type": "string",
            "description": "Id креативной площадки",
            "name": "id",
            "in": "path",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "allOf": [
                {
                  "$ref": "#/definitions/model.ResponseSuccess"
                },
                {
                  "type": "object",
                  "properties": {
                    "data": {
                      "$ref": "#/definitions/responses.getCreativeSpaceResponseData"
                    }
                  }
                }
              ]
            }
          },
          "403": {
            "description": "Коды ошибок: [1100]",
            "schema": {
              "$ref": "#/definitions/model.ResponseError"
            }
          }
        }
      },
      "delete": {
        "consumes": ["application/json"],
        "produces": ["application/json"],
        "summary": "Удаление информации о креативной площадке",
        "parameters": [
          {
            "type": "string",
            "description": "Id креативной площадки",
            "name": "id",
            "in": "path",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/model.ResponseSuccess"
            }
          },
          "403": {
            "description": "Коды ошибок: [1100]",
            "schema": {
              "$ref": "#/definitions/model.ResponseError"
            }
          }
        }
      },
      "patch": {
        "consumes": ["application/json"],
        "produces": ["application/json"],
        "summary": "Обновление информации о креативном пространстве",
        "parameters": [
          {
            "type": "string",
            "description": "Id креативного пространства",
            "name": "id",
            "in": "path",
            "required": true
          },
          {
            "description": "Параметры запроса",
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/requests.patchCreativeSpaceRequestData"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/model.ResponseSuccess"
            }
          },
          "403": {
            "description": "Коды ошибок: [1100]",
            "schema": {
              "$ref": "#/definitions/model.ResponseError"
            }
          }
        }
      }
    },
    "/api/v1/file": {
      "post": {
        "consumes": ["multipart/form-data"],
        "produces": ["application/json"],
        "summary": "Возвращает ссылку на файл",
        "parameters": [
          {
            "type": "string",
            "description": "Authentication header",
            "name": "Authorization",
            "in": "header",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "allOf": [
                {
                  "$ref": "#/definitions/model.ResponseSuccess"
                },
                {
                  "type": "object",
                  "properties": {
                    "data": {
                      "$ref": "#/definitions/responses.createFileStationsResponseData"
                    }
                  }
                }
              ]
            }
          },
          "403": {
            "description": "Коды ошибок: [1100]",
            "schema": {
              "$ref": "#/definitions/model.ResponseError"
            }
          }
        }
      }
    },
    "/api/v1/files/:fileName": {
      "delete": {
        "consumes": ["application/json"],
        "summary": "Удаляет файл из s3",
        "parameters": [
          {
            "type": "string",
            "description": "Authentication header",
            "name": "Authorization",
            "in": "header",
            "required": true
          },
          {
            "description": "Параметры запроса",
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/requests.deleteFileRequestData"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/model.ResponseSuccess"
            }
          },
          "403": {
            "description": "Коды ошибок: [1100]",
            "schema": {
              "$ref": "#/definitions/model.ResponseError"
            }
          }
        }
      }
    },
...
    "/api/v1/metroStations": {
      "get": {
        "produces": ["application/json"],
        "summary": "Возвращает полный список станций метро",
        "parameters": [
          {
            "type": "string",
            "description": "Authentication header",
            "name": "Authorization",
            "in": "header",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "allOf": [
                {
                  "$ref": "#/definitions/model.ResponseSuccess"
                },
                {
                  "type": "object",
                  "properties": {
                    "data": {
                      "$ref": "#/definitions/responses.getMetroStationsResponseData"
                    }
                  }
                }
              ]
            }
          },
          "403": {
            "description": "Коды ошибок: [1100]",
            "schema": {
              "$ref": "#/definitions/model.ResponseError"
            }
          }
        }
      }
    },
    "/api/v1/score": {
      "post": {
        "consumes": ["application/json"],
        "produces": ["application/json"],
        "summary": "Бронирование креативной площадки",
        "parameters": [
          {
            "description": "Параметры запроса",
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/requests.createScoreRequestData"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "allOf": [
                {
                  "$ref": "#/definitions/model.ResponseSuccess"
                },
                {
                  "type": "object",
                  "properties": {
                    "data": {
                      "$ref": "#/definitions/responses.createScoreResponseData"
                    }
                  }
                }
              ]
            }
          },
          "403": {
            "description": "Коды ошибок: [1100]",
            "schema": {
              "$ref": "#/definitions/model.ResponseError"
            }
          }
        }
      }
    },
    "/api/v1/session": {
      "get": {
        "summary": "Получение сессии",
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/model.ResponseSuccess"
            }
          },
          "403": {
            "description": "Коды ошибок: [1100]",
            "schema": {
              "$ref": "#/definitions/model.ResponseError"
            }
          }
        }
      },
      "post": {
        "consumes": ["application/json"],
        "produces": ["application/json"],
        "summary": "Создание сессии пользователя",
        "parameters": [
          {
            "description": "Параметры запроса",
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/requests.createSessionRequestData"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "allOf": [
                {
                  "$ref": "#/definitions/model.ResponseSuccess"
                },
                {
                  "type": "object",
                  "properties": {
                    "data": {
                      "$ref": "#/definitions/responses.createSessionResponseData"
                    }
                  }
                }
              ]
            }
          }
        }
      },
      "delete": {
        "summary": "Удаление сессии",
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/model.ResponseSuccess"
            }
          },
          "403": {
            "description": "Коды ошибок: [1100]",
            "schema": {
              "$ref": "#/definitions/model.ResponseError"
            }
          }
        }
      }
    },
    "/api/v1/user": {
      "post": {
        "consumes": ["application/json"],
        "produces": ["application/json"],
        "summary": "Создание пользователя",
        "parameters": [
          {
            "description": "Параметры запроса",
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/requests.createUserRequestData"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "allOf": [
                {
                  "$ref": "#/definitions/model.ResponseSuccess"
                },
                {
                  "type": "object",
                  "properties": {
                    "data": {
                      "$ref": "#/definitions/responses.createUserResponseData"
                    }
                  }
                }
              ]
            }
          },
          "403": {
            "description": "Коды ошибок: [1100]",
            "schema": {
              "$ref": "#/definitions/model.ResponseError"
            }
          }
        }
      }
    },
    "/api/v1/users": {
      "get": {
        "consumes": ["application/json"],
        "produces": ["application/json"],
        "summary": "Возвращает список пользователей",
        "parameters": [
          {
            "type": "string",
            "description": "Authentication header",
            "name": "Authorization",
            "in": "header",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "allOf": [
                {
                  "$ref": "#/definitions/model.ResponseSuccess"
                },
                {
                  "type": "object",
                  "properties": {
                    "data": {
                      "$ref": "#/definitions/responses.getUsersResponseData"
                    }
                  }
                }
              ]
            }
          },
          "403": {
            "description": "Коды ошибок: [1100]",
            "schema": {
              "$ref": "#/definitions/model.ResponseError"
            }
          }
        }
      }
    },
    "/api/v1/users/:id": {
      "get": {
        "produces": ["application/json"],
        "summary": "Возвращает информацию о пользователе",
        "parameters": [
          {
            "type": "string",
            "description": "Id пользователя",
            "name": "id",
            "in": "path",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "allOf": [
                {
                  "$ref": "#/definitions/model.ResponseSuccess"
                },
                {
                  "type": "object",
                  "properties": {
                    "data": {
                      "$ref": "#/definitions/responses.getUserResponseData"
                    }
                  }
                }
              ]
            }
          },
          "403": {
            "description": "Коды ошибок: [1100]",
            "schema": {
              "$ref": "#/definitions/model.ResponseError"
            }
          }
        }
      },
      "delete": {
        "consumes": ["application/json"],
        "produces": ["application/json"],
        "summary": "Удаление информации о пользователе",
        "parameters": [
          {
            "type": "string",
            "description": "Id пользователя",
            "name": "id",
            "in": "path",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/model.ResponseSuccess"
            }
          },
          "403": {
            "description": "Коды ошибок: [1100]",
            "schema": {
              "$ref": "#/definitions/model.ResponseError"
            }
          }
        }
      },
      "patch": {
        "consumes": ["application/json"],
        "produces": ["application/json"],
        "summary": "Обновление информации о пользователе",
        "parameters": [
          {
            "type": "string",
            "description": "Id пользователя",
            "name": "id",
            "in": "path",
            "required": true
          },
          {
            "description": "Параметры запроса",
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/requests.patchUserRequestData"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/model.ResponseSuccess"
            }
          },
          "403": {
            "description": "Коды ошибок: [1100]",
            "schema": {
              "$ref": "#/definitions/model.ResponseError"
            }
          }
        }
      }
    }
  },
  "definitions": {
    "model.ResponseError": {
      "type": "object",
      "properties": {
        "error": {
          "$ref": "#/definitions/model.ResponseErrorField"
        },
        "success": {
          "type": "boolean",
          "default": false
        }
      }
    },
...
    "model.ResponseSuccess": {
      "type": "object",
      "properties": {
        "data": {},
        "success": {
          "type": "boolean",
          "default": true
        }
      }
    },
...
    "requests.createBookingRequestData": {
      "type": "object",
      "required": ["calendarEvents", "creativeSpaceId"],
      "properties": {
        "calendarEvents": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/requests.createBookingRequestCalendarEvent"
          }
        },
        "creativeSpaceId": {
          "type": "integer"
        }
      }
    },
...
    "requests.createCreativeSpaceRequestData": {
      "type": "object",
      "required": [
        "address",
        "area",
        "calendar",
        "capacity",
        "coordinate",
        "description",
        "photos",
        "pricePerDay",
        "spaceType",
        "title"
      ],
      "properties": {
        "address": {
          "type": "string"
        },
        "area": {
          "type": "integer"
        },
        "calendar": {
          "$ref": "#/definitions/requests.createCreativeSpaceRequestCalendar"
        },
        "capacity": {
          "type": "integer"
        },
        "coordinate": {
          "$ref": "#/definitions/requests.createCreativeSpaceRequestCoordinate"
        },
        "description": {
          "type": "string"
        },
        "metroStations": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/requests.createCreativeSpaceRequestMetroStation"
          }
        },
        "photos": {
          "type": "array",
          "items": {
            "type": "string"
          }
        },
        "pricePerDay": {
          "type": "integer"
        },
        "spaceType": {
          "type": "string"
        },
        "title": {
          "type": "string"
        }
      }
    },
...
    "requests.patchBookingRequestData": {
      "type": "object",
      "properties": {
        "calendarEvents": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/requests.patchBookingRequestCalendarEvent"
          }
        },
        "id": {
          "type": "integer"
        },
        "status": {
          "type": "string"
        }
      }
    },
...
    "responses.convertCalendarFromLinkResponseData": {
      "type": "object",
      "properties": {
        "events": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/responses.convertCaledarFromLinkResponseEvent"
          }
        }
      }
    },
...
    "responses.convertCalendarResponseData": {
      "type": "object",
      "properties": {
        "events": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/responses.convertCaledarResponseEvent"
          }
        }
      }
    },
...
    "responses.createBookingResponseData": {
      "type": "object",
      "properties": {
        "booking": {
          "$ref": "#/definitions/responses.createBookingResponseBooking"
        }
      }
    },
...
    "responses.createCreativeSpaceResponseData": {
      "type": "object",
      "properties": {
        "creativeSpace": {
          "$ref": "#/definitions/responses.createCreativeSpaceResponseCreativeSpace"
        }
      }
    },
...
    "responses.createFileStationsResponseData": {
      "type": "object",
      "properties": {
        "file": {
          "$ref": "#/definitions/responses.createFileResponseFile"
        }
      }
    },
...
    "responses.getAddressesResponseData": {
      "type": "object",
      "properties": {
        "addresses": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/responses.getAddressesResponseUser"
          }
        }
      }
    },
...
    "responses.getBookingResponseData": {
      "type": "object",
      "properties": {
        "booking": {
          "$ref": "#/definitions/responses.getBookingResponseBooking"
        }
      }
    },
...
    "responses.getBookingsResponseData": {
      "type": "object",
      "properties": {
        "bookings": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/responses.getBookingsResponseBooking"
          }
        }
      }
    },
...
    "responses.getCreativeSpaceResponseData": {
      "type": "object",
      "properties": {
        "creativeSpace": {
          "$ref": "#/definitions/responses.getCreativeSpaceResponseCreativeSpace"
        }
      }
    },
...
    "responses.getCreativeSpacesResponseData": {
      "type": "object",
      "properties": {
        "creativeSpaces": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/responses.getCreativeSpacesResponseCreativeSpace"
          }
        }
      }
    },
...
  }
}